+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## v1.2.0
+### 2021-11-18
+
+The `VariableBase` class has been updated to include an `is_temporal` method.
+This will  return a boolean value indicating whether the variable has a `units`
+metadata attribute conforming to the format outlined in the CF-Conventions
+(section 4.4). A convenience method `Variable.get_attribute_value` has been
+added This will allow an end-user to retrieve the value of a metadata
+attribute, while specifying a default value. Lastly, the
+`VarInfoBase.get_temporal_dimensions` method has been added, which retrieves a
+set of required `Variables` for which `Variable.is_temporal()` is true, based
+on a set of requested variables.
+
 ## v1.1.0
 ### 2021-09-03
 
@@ -15,7 +28,7 @@ method.
 
 The `VarInfoBase.get_required_variables` and
 `VarInfoBase.get_required_dimensions` methods have been updated to filter out
-any variable references that point to non-existant variables. This might occur
+any variable references that point to non-existent variables. This might occur
 when a variable has a dimension only present to denote array size. For example,
 a CF-Convention `bounds` reference variable, such as `lat_bnds`. In these
 cases, the variable will have an additional dimension without a corresponding